Integrate Canadian and American licen[cs]e
I’m Canadian and spell “licence” the Canadian way, thus the document
refers to “licence” and “licences”. However, the licences that are used
are properly titled using the American spelling “license”. Just as the
Toronto Maple Leafs play in the “Air Canada Centre” and the Buffalo
Sabres play in the “First Niagara Center”, it is correct to refer to the
“MIT License” with the American spelling.

To complicate things further, the document permitting the bearer to
drive is called a “licence” (noun). The process by which one gets that
document is “licensing” (verb).

This commit fixes places where I had mixed the spellings to standardize
the non-titled version on the Canadian spelling and the titled version
as appropriate to the licence in question.

Thanks to @davidcelis for bringing this up in PR #54.