Calculate priority from MIME types of same family
Some MIME types are in the same family, but their #simplified values
aren't the same. One such example are "text/comma-separated-values" and
"text/csv" MIME types. The former is obsolete and unregistered, while
the latter is not obsolete and registered, but mime-types would still
rank "text/comma-separated-values" higher because it just looks at the
alphabetic order.